Move most remaining arithmetic tests
