+++ pandas/tests/indexes/period/test_ops.py
@@ -412,94 +412,3 @@ class TestPeriodIndexOps(Ops):
         # warning for setter
         with tm.assert_produces_warning(FutureWarning):
             idx.freq = pd.offsets.Day()
-
-
-class TestPeriodIndexSeriesMethods(object):
-    """ Test PeriodIndex and Period Series Ops consistency """
-
-    def _check(self, values, func, expected):
-        idx = pd.PeriodIndex(values)
-        result = func(idx)
-        if isinstance(expected, pd.Index):
-            tm.assert_index_equal(result, expected)
-        else:
-            # comp op results in bool
-            tm.assert_numpy_array_equal(result, expected)
-
-        s = pd.Series(values)
-        result = func(s)
-
-        exp = pd.Series(expected, name=values.name)
-        tm.assert_series_equal(result, exp)
-
-    def test_pi_comp_period(self):
-        idx = PeriodIndex(['2011-01', '2011-02', '2011-03',
-                           '2011-04'], freq='M', name='idx')
-
-        f = lambda x: x == pd.Period('2011-03', freq='M')
-        exp = np.array([False, False, True, False], dtype=np.bool)
-        self._check(idx, f, exp)
-        f = lambda x: pd.Period('2011-03', freq='M') == x
-        self._check(idx, f, exp)
-
-        f = lambda x: x != pd.Period('2011-03', freq='M')
-        exp = np.array([True, True, False, True], dtype=np.bool)
-        self._check(idx, f, exp)
-        f = lambda x: pd.Period('2011-03', freq='M') != x
-        self._check(idx, f, exp)
-
-        f = lambda x: pd.Period('2011-03', freq='M') >= x
-        exp = np.array([True, True, True, False], dtype=np.bool)
-        self._check(idx, f, exp)
-
-        f = lambda x: x > pd.Period('2011-03', freq='M')
-        exp = np.array([False, False, False, True], dtype=np.bool)
-        self._check(idx, f, exp)
-
-        f = lambda x: pd.Period('2011-03', freq='M') >= x
-        exp = np.array([True, True, True, False], dtype=np.bool)
-        self._check(idx, f, exp)
-
-    def test_pi_comp_period_nat(self):
-        idx = PeriodIndex(['2011-01', 'NaT', '2011-03',
-                           '2011-04'], freq='M', name='idx')
-
-        f = lambda x: x == pd.Period('2011-03', freq='M')
-        exp = np.array([False, False, True, False], dtype=np.bool)
-        self._check(idx, f, exp)
-        f = lambda x: pd.Period('2011-03', freq='M') == x
-        self._check(idx, f, exp)
-
-        f = lambda x: x == tslib.NaT
-        exp = np.array([False, False, False, False], dtype=np.bool)
-        self._check(idx, f, exp)
-        f = lambda x: tslib.NaT == x
-        self._check(idx, f, exp)
-
-        f = lambda x: x != pd.Period('2011-03', freq='M')
-        exp = np.array([True, True, False, True], dtype=np.bool)
-        self._check(idx, f, exp)
-        f = lambda x: pd.Period('2011-03', freq='M') != x
-        self._check(idx, f, exp)
-
-        f = lambda x: x != tslib.NaT
-        exp = np.array([True, True, True, True], dtype=np.bool)
-        self._check(idx, f, exp)
-        f = lambda x: tslib.NaT != x
-        self._check(idx, f, exp)
-
-        f = lambda x: pd.Period('2011-03', freq='M') >= x
-        exp = np.array([True, False, True, False], dtype=np.bool)
-        self._check(idx, f, exp)
-
-        f = lambda x: x < pd.Period('2011-03', freq='M')
-        exp = np.array([True, False, False, False], dtype=np.bool)
-        self._check(idx, f, exp)
-
-        f = lambda x: x > tslib.NaT
-        exp = np.array([False, False, False, False], dtype=np.bool)
-        self._check(idx, f, exp)
-
-        f = lambda x: tslib.NaT >= x
-        exp = np.array([False, False, False, False], dtype=np.bool)
-        self._check(idx, f, exp)
