+++ pandas/tests/series/test_arithmetic.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 import operator
 
-import numpy as np
 import pytest
 
 from pandas import Series
@@ -14,13 +13,6 @@ import pandas.util.testing as tm
 # Comparisons
 
 class TestSeriesComparison(object):
-    def test_compare_invalid(self):
-        # GH#8058
-        # ops testing
-        a = pd.Series(np.random.randn(5), name=0)
-        b = pd.Series(np.random.randn(5))
-        b.name = pd.Timestamp('2000-01-01')
-        tm.assert_series_equal(a / b, 1 / (b / a))
 
     @pytest.mark.parametrize('opname', ['eq', 'ne', 'gt', 'lt', 'ge', 'le'])
     def test_ser_flex_cmp_return_dtypes(self, opname):
